RestApi CRUD 완료

+++ src/main/java/site/metacoding/white/service/BoardService.java
@@ -1,6 +1,8 @@
 package site.metacoding.white.service;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -10,12 +12,13 @@ import site.metacoding.white.domain.Board;
 import site.metacoding.white.domain.BoardRepository;
 import site.metacoding.white.dto.BoardReqDto.BoardSaveReqDto;
 import site.metacoding.white.dto.BoardReqDto.BoardUpdateReqDto;
+import site.metacoding.white.dto.BoardRespDto.BoardDetailRespDto;
+import site.metacoding.white.dto.BoardRespDto.BoardDetailRespDto.BoardAllRespDto;
 import site.metacoding.white.dto.BoardRespDto.BoardSaveRespDto;
 import site.metacoding.white.dto.BoardRespDto.BoardUpdateRespDto;
 
 // 트랜젝션 관리
 // DTO 변환하여 컨트롤러에게 리턴
-
 @RequiredArgsConstructor
 @Service // ioc컨테이너에 띄워줌
 public class BoardService {
@@ -34,24 +37,52 @@ public class BoardService {
 		return boardSaveRespDto;
 	}
 
-	public Board findById(Long id) { // select만 할거면 transactional 안붙여도 됨
-		return boardRepository.findById(id);
+	public BoardDetailRespDto findById(Long id) {
+		Optional<Board> boardOP = boardRepository.findById(id);
+		if (boardOP.isPresent()) {
+			BoardDetailRespDto boardDetailRespDto = new BoardDetailRespDto(boardOP.get());
+			return boardDetailRespDto;
+		} else {
+			throw new RuntimeException("해당" + id + "로 상세보기를 할 수 없습니다.");
+		}
 	}
 
-	public List<Board> findAll() {
-		return boardRepository.findAll();
+	public List<BoardAllRespDto> findAll() {
+		List<Board> boardList = boardRepository.findAll();
+
+		List<BoardAllRespDto> boardAllRespDtoList = new ArrayList<>();
+		// 1. for문 돌리기(list의 크기만큼)
+		for (Board board : boardList) { // Board 변수선언(board) : 배열명
+			boardAllRespDtoList.add(new BoardAllRespDto(board));
+		}
+		return boardAllRespDtoList;
+		// 2. Board -> Dto로 옮기기
+		// 3. Dto를 list에 담기
 	}
 
 	@Transactional
-	public BoardUpdateRespDto update(Long id, BoardUpdateReqDto boardUpdateReqDto) {
-		Board boardPS = boardRepository.findById(id);
-		boardPS.update(boardUpdateReqDto.getTitle(), boardUpdateReqDto.getContent());
-		return new BoardUpdateRespDto(boardPS);
+	public BoardUpdateRespDto update(BoardUpdateReqDto boardUpdateReqDto) {
+		Long id = boardUpdateReqDto.getId();
+		Optional<Board> boardOP = boardRepository.findById(boardUpdateReqDto.getId());
+		if (boardOP.isPresent()) {
+			Board boardPS = boardOP.get();
+			boardPS.update(boardUpdateReqDto.getTitle(), boardUpdateReqDto.getContent());
+			return new BoardUpdateRespDto(boardOP.get());
+		} else {
+			throw new RuntimeException("해당" + id + "로 업데이트를 할 수 없습니다.");
+		}
+
 	} // 종료시 더티체킹 하여 모든 쓰레기 데이터를 flush => update 됨
 
+	// delete는 리턴 안함
 	@Transactional
 	public void deleteById(Long id) {
-		boardRepository.deleteById(id);
+		Optional<Board> boardOP = boardRepository.findById(id);
+		if (boardOP.isPresent()) {
+			boardRepository.deleteById(id);
+		} else {
+			throw new RuntimeException("해당" + id + "로 삭제를 할 수 없습니다.");
+		}
 	}
 
 }
